requirements.txt
requests==2.31.0
pytz==2024.1
geopy==2.4.1
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
tensorflow
flask==3.0.1
scikit-learn==1.4.2
gunicorn==22.0.0
s3fs==2023.12.2